refactor(ui): move 5 pure modals to dbflux_components::modals (#102)
Move active_query, drop_table, delete_connection, tunnel_auth, schema_drift
from dbflux_ui into dbflux_components::modals/. Rewrite imports to use
crate:: intra-crate paths (no dbflux_components:: refs inside the crate).
Replace each source file in dbflux_ui with a pub-use shim. Tests are
relocated alongside the code they verify (drop_table, tunnel_auth).
