+++ crates/dbflux_components/src/modals/mod.rs
@@ -1,3 +1,13 @@
+pub mod active_query;
+pub mod delete_connection;
+pub mod drop_table;
+pub mod schema_drift;
 pub mod shell;
+pub mod tunnel_auth;
 
+pub use active_query::{ActiveQueryOutcome, ActiveQueryRequest, ActiveQueryTrigger, ModalActiveQuery};
+pub use delete_connection::{DeleteConnectionOutcome, DeleteConnectionRequest, ModalDeleteConnection};
+pub use drop_table::{DropTableOutcome, DropTableRequest, ModalDropTable};
+pub use schema_drift::{ModalSchemaDrift, SchemaDriftContinue, SchemaDriftDismissed, SchemaDriftRefresh};
 pub use shell::{ModalShell, ModalVariant};
+pub use tunnel_auth::{ModalTunnelAuth, TunnelAuthOutcome, TunnelAuthRequest};
